+++ unicorn/provider/com/github/gist.lua
@@ -1,9 +1,11 @@
 package.path = "/lib/?.lua;/lib/?;/lib/?/init.lua;" .. package.path
 local unicorn = require("unicorn")
 
+local provider = {}
+
 -- Package provider for GitHub Gists <gists.github.com>.
 ---@param package_table table A valid package table
-local function install_gist(package_table)
+function provider.install(package_table)
 	-- this is really simple, only works predictably with a one-file gist.
 	for remote_path, install_path in pairs(package_table.instdat.filemaps) do
 		-- uses source code repository-like names because a Gist is a repository technically :)
@@ -19,4 +21,4 @@ local function install_gist(package_table)
 	end
 end
 
-return install_gist
+return provider